@@ -27,45 +27,49 @@
     "dev": "NODE_ENV=development      node --harmony src/scripts/make.js  &&  NODE_ENV=development  ./node_modules/nodemon/bin/nodemon.js --harmony src/server.js",
     "make": "NODE_ENV=development     node --harmony src/scripts/make.js",
     "lint": "eslint . --ext .js --ext .jsx",
+    "lint-fix": "eslint . --ext .js --ext .jsx --fix",
     "postinstall": "npm run build",
     "test": "./node_modules/.bin/mocha --compilers jsx:babel-register --require ignore-styles ./tests/specs --recursive"
   },
   "dependencies": {
-    "assets-webpack-plugin": "^3.2.0",
-    "babel-core": "^6.3.17",
-    "babel-eslint": "^4.1.6",
-    "babel-loader": "^6.2.0",
-    "babel-plugin-react-transform": "^2.0.0-beta1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-1": "^6.3.13",
-    "babel-register": "^6.3.13",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.1",
-    "css-loader": "^0.23.0",
-    "hapi": "^11.1.2",
-    "hapi-react-views": "^5.0.1",
-    "ignore-styles": "^1.2.0",
-    "inert": "^3.2.0",
-    "jsdom": "^8.1.0",
-    "mocha": "^2.4.5",
-    "react": "^0.14.3",
-    "react-addons-test-utils": "^0.14.7",
-    "react-dom": "^0.14.3",
-    "rimraf": "^2.4.4",
-    "style-loader": "^0.13.0",
-    "vision": "^4.0.1",
-    "webpack": "^1.11.0",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.3.0"
+    "assets-webpack-plugin": "3.5.1",
+    "babel-core": "6.24.1",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "6.4.1",
+    "babel-plugin-react-transform": "2.0.2",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-1": "6.24.1",
+    "babel-register": "6.24.1",
+    "chai": "3.5.0",
+    "chalk": "1.1.3",
+    "css-loader": "0.28.0",
+    "hapi": "16.1.1",
+    "hapi-react-views": "9.2.1",
+    "ignore-styles": "5.0.1",
+    "inert": "4.2.0",
+    "jsdom": "9.12.0",
+    "mocha": "3.2.0",
+    "prop-types": "15.5.8",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
+    "rimraf": "2.6.1",
+    "style-loader": "0.16.1",
+    "vision": "4.1.1",
+    "webpack": "1.15.0",
+    "webpack-dev-middleware": "1.10.2",
+    "webpack-hot-middleware": "2.18.0"
   },
   "devDependencies": {
-    "eslint": "^1.1.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^3.2.2",
-    "express": "^4.13.3",
-    "h2o2": "^4.0.1",
-    "nodemon": "^1.8.1",
-    "react-transform-hmr": "^1.0.1"
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "14.1.0",
+    "eslint-loader": "1.7.1",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "4.0.0",
+    "eslint-plugin-react": "6.10.3",
+    "express": "4.15.2",
+    "h2o2": "5.4.0",
+    "nodemon": "1.11.0",
+    "react-transform-hmr": "1.0.4"
   }
 }
