@@ -1,61 +1,65 @@
-'use strict';
+const React = require('react');
+const PropTypes = require('prop-types');
 
-var React = require('react');
 require('./counter.css');
 
-/**
- * @class Counter
- * @extends ReactComponent
- */
 class Counter extends React.Component {
 
-    constructor (props) {
-        super(props);
+  constructor(props) {
+    super(props);
 
-        this.state = {
-            counter: 0
-        };
+    this.state = {
+      counter: 0
+    };
 
-        this.increment = this.increment.bind(this);
-        this.remove = this.remove.bind(this);
+    this.increment = this.increment.bind(this);
+    this.remove = this.remove.bind(this);
 
-        setInterval(this.increment.bind(this), this.props.interval);
-    }
+    setInterval(this.increment, this.props.interval);
+  }
 
-    render () {
-        return (
-            <div className='Counter'>
+  increment() {
+    this.setState({ counter: this.state.counter + 1 });
+  }
 
-                <h1>Counter</h1>
+  remove() {
+    this.setState({ counter: this.state.counter - 1 });
+  }
 
-                <p>This is a counter component. It stores the count in its state.</p>
-                <p>If you edit any file that is bundled by webpack, webpack dev server will patch this page while preserving component state.</p>
-                <p>This is called <em>hot module reloading</em> and is more useful than just auto-reloading the whole page (if you reload the page, you'll see the counter drops to zero because component state is not persisted).</p>
+  render() {
+    return (
+      <div className="Counter">
 
-                <span className='-number'>{this.state.counter}</span>
-                <button className='-button' type='button' onClick={this.increment.bind(this)}>+</button>
-                <button className='-button' type='button' onClick={this.remove.bind(this)}>–</button>
+        <h1>Counter</h1>
 
-            </div>
-        );
-    }
+        <p>This component is rendered on the <strong>frontend</strong>.</p>
+        <p>This is a counter component. It stores the count in its state.</p>
+        <p>
+          If you edit any file that is bundled by webpack, webpack dev server will
+          patch this page while preserving component state, without reloading the page.
+        </p>
+        <p>
+          This is called <em>hot module reloading</em> and is more useful than just
+          auto-reloading the whole page (if you reload the page, you&apos;ll see the counter
+          drops to zero because component state is not persisted).
+        </p>
 
-    increment () {
-        super.setState({counter: this.state.counter + 1});
-    }
+        <span className="-number">{this.state.counter}</span>
+        <button className="-button" type="button" onClick={this.increment}>+</button>
+        <button className="-button" type="button" onClick={this.remove}>–</button>
 
-    remove () {
-        super.setState({counter: this.state.counter - 1});
-    }
+      </div>
+    );
+  }
 }
 
 
 Counter.defaultProps = {
-    interval: 1000
+  interval: 1000
 };
 
 Counter.propTypes = {
-    interval: React.PropTypes.number.isRequired
+  interval: PropTypes.number.isRequired
 };
 
 
