@@ -1,20 +1,18 @@
-'use strict';
-
-var path = require('path');
-var FileWriter = require('../tools/file-writer');
-var config = require('../config/variables');
+const path = require('path');
+const FileWriter = require('../tools/file-writer');
+const config = require('../config/variables');
 
 
-var eslintIgnore = [
-    '###',
-    '### WARNING: This file is generated by ' + path.basename(__filename) + ' - do not edit manually!',
-    '###',
-    '',
-    path.relative(config.paths.root, config.paths.build),
-    ''
+const eslintIgnore = [
+  '###',
+  `### WARNING: This file is generated by ${path.basename(__filename)} - do not edit manually!`,
+  '###',
+  '',
+  path.relative(config.paths.root, config.paths.build),
+  ''
 ];
 
 
-module.exports = function () {
-    FileWriter.write('.eslintignore', eslintIgnore.join('\n'));
+module.exports = function makeEslintIgnore() {
+  FileWriter.write('.eslintignore', eslintIgnore.join('\n'));
 };
