@@ -1,38 +1,41 @@
-'use strict';
+/* eslint no-console: ["off"] */
 
-var chalk = require('chalk');
+const chalk = require('chalk');
 
-var logPrefix = '[WebpackSlowPlugin]: ';
+const logPrefix = '[WebpackSlowPlugin]: ';
 
 
-function SlowWebpackPlugin (options) {
-    options = options || {};
-    options.delay = parseInt(options.delay);
-    options.delay = !isNaN(options.delay) ? options.delay : 1000;
+function SlowWebpackPlugin(options) {
+  const delay = parseInt(options.delay, 10);
+  const cleanOptions = Object.assign(
+    {},
+    options,
+    { delay: !isNaN(delay) ? delay : 1000 }
+  );
 
-    this.options = options;
+  this.options = cleanOptions;
 }
 
 
-SlowWebpackPlugin.prototype.apply = function (compiler) {
-    var delay = this.options.delay;
+SlowWebpackPlugin.prototype.apply = function apply(compiler) {
+  const delay = this.options.delay;
 
-    compiler.plugin('done', () => {
-        var beginTime = Date.now();
-        var curTime = beginTime;
-        var secondsElapsed = 0;
-        console.log('');
-        console.log(chalk.yellow(logPrefix + 'Begin'));
-        while (curTime - beginTime < delay) {
-            curTime = Date.now();
-            if (Math.floor((curTime - beginTime) / 1000) > secondsElapsed) {
-                secondsElapsed++;
-                console.log(chalk.yellow(logPrefix + secondsElapsed + '/' + Math.ceil(delay / 1000)));
-            }
-        }
-        console.log(chalk.yellow(logPrefix + 'End'));
-        console.log('');
-    });
+  compiler.plugin('done', () => {
+    const beginTime = Date.now();
+    let curTime = beginTime;
+    let secondsElapsed = 0;
+    console.log('');
+    console.log(chalk.yellow(`${logPrefix}Begin`));
+    while (curTime - beginTime < delay) {
+      curTime = Date.now();
+      if (Math.floor((curTime - beginTime) / 1000) > secondsElapsed) {
+        secondsElapsed += 1;
+        console.log(chalk.yellow(`${logPrefix + secondsElapsed}/${Math.ceil(delay / 1000)}`));
+      }
+    }
+    console.log(chalk.yellow(`${logPrefix}End`));
+    console.log('');
+  });
 };
 
 
